fix: PR #532レビュー指摘を修正(ラベル衝突ハンドリング・レベル採番・テスト網羅)
12角度レビュー+敵対的再レビューで確定した所見を全件修正。

API/シード/スキーマ:
- 共有ヘルパー lib/db-errors.ts の isLabelConflictError を新設し、0041
  BEFOREトリガーの中断メッセージ(...label already exists)とユニーク
  インデックスの UNIQUE constraint failed の両形を一致。game-group/
  variant/mix の create/update とセルフシードで一元化(旧
  isUniqueConstraintViolation を廃止)。参照整合トリガーは非一致を担保。
- seed-game-data: 並行シード敗者のトリガー中断のみ握りつぶし(no-op)、
  非ラベルエラーは再throw。ビルトインラベル相互一意の不変条件テストで固定。
- session.update: タグ置換を単一 runBatch で原子化(DELETE+再INSERTを
  同一バッチ、SA2-116)。原子性/10列チャンク境界テストを強化。
- live-cash-game-session: initialBuyIn を .int().min(0) に(read=write、SA2-148)。

Web:
- 新規レベル番号を max(level)+1 に統一。共有純関数 nextLevelNumber に
  一元化し、use-blind-levels とローカルエディタ helpers の両経路が同一規則で
  採番(gappyなサーバデータでの重複番号を解消)。
- デッドコード削除: useTournaments の未使用 *WithLevels ラッパー(games を
  落とす)/重複 use-empty-game-set-rows(blind3 欠落)を削除し、実フックの
  テストを新設。
- mix-form-sheet: 新規バリアントの id 解決を gameVariant.list キャッシュに
  フォールバック(無言脱落を防止)。
- games-page: variantLabelById 導出をフックへ移し MixesCard を純表示化。
  全id未解決時の "N games" フォールバックを網羅。
- game-scene-formatters: anteType を ?? "all" に(null≈undefined 正規化)。

検証: ultracite クリーン / check-types(server+web)緑 / vitest --changed
HEAD 1,936 passed。

+++ packages/api/src/routers/live-cash-game-session.ts
@@ -413,7 +413,11 @@ export const liveCashGameSessionRouter = router({
 				ringGameId: z.string().optional(),
 				currencyId: z.string().optional(),
 				memo: z.string().optional(),
-				initialBuyIn: z.number().min(0),
+				// Must be an integer: it is re-read through cashSessionStartPayload's
+				// z.number().int().min(0), so a decimal here would parse on create but
+				// throw ZodError on every subsequent getById, making the session
+				// permanently unreadable (SA2-148 write==read).
+				initialBuyIn: z.number().int().min(0),
 			})
 		)
 		.mutation(async ({ ctx, input }) => {
